Added some tests that check a complete event/transaction json to be same in SDK 1.x and 2.x
